move mesh class
